// src/app/extends/extend-btn-names.enum.ts
export enum ExtendBtnNames {
  loginPageBtnName = 'الدخول',
  resetPagePasswordBtnName = '',
  companyRegisterBtnName = 'التسجيل كاشركة',
  ConfirmationPageBtnName = '',
  createAdPageBtnName = '',
  registerPageBtnName = 'تسجيل',
  completeRegisterPageName = '',
  saveBtnName = '',
  searchBtn = 'بحث',
  addAnotherProjectBtnName = '',
  readMoreBtnName = 'اقرأ المزيد',
  commentBtnName = '',
  purchasePageBtnName = '',
  fooFormBtn = 'إرسـال',
  editBtn = 'تعديل',
  deleteBtn = 'مسح',
  addProjectBtn = 'اضافة المشروع',
  addAdbtn = 'اضافة الاعلان',
  updateBtn = 'تحديث',
  forgetPass ='تسيت كلمة السر'
}