// src/app/extends/extend-form-labels.enum.ts
export enum ExtendLabels {
  nameLabel = 'الأسم',
  emailLabel = 'الإيميل',
  messageLabel = 'الرسالة',
  passwordLabel = 'كلمة السر',
  oldpasswordLabel = 'كلمة السر القديمة',
  newpasswordLabel = 'كلمة السر الجديدة',
  renewpasswordLabel = 'تأكيد كلمة السر الجديدة',
  sendEmailAddres = '',
  sendPhoneMessage = '',
  phoneNumLabel = 'رقم الهاتف',
  countryLabel = 'أختر الدولة',
  cityLabel = 'أختر المدينه',
  companyCat = 'فئة الشركه',
  companyIndustry = 'تخصص الشركة',
  commentLabel = '',
  cardNumberLabel = '',
  cardCodeLabel = '',
  ExpireDate = '',
  footerMessageLabel = ''

}